+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Create new layout for My Returns List and Request Returns Available List (change from table to Grid Layout with cards)
+### Fixed
+- Change structure of solicitation details and new request (from table to card) in mobile devices.
 
 ## [3.5.6] - 2023-02-06